HBASE-2402 [stargate] set maxVersions on gets HBASE-2412 [stargate] PerformanceEvaluation HBASE-2319 [stargate] multiuser mode: request shaping HBASE-2403 [stargate] client HTable interface to REST connector
git-svn-id: https://svn.apache.org/repos/asf/hadoop/hbase/branches/0.20_pre_durability@932214 13f79535-47bb-0310-9956-ffa450edef68

+++ src/contrib/stargate/src/java/org/apache/hadoop/hbase/stargate/SchemaResource.java
@@ -35,10 +35,12 @@ import javax.ws.rs.core.Context;
 import javax.ws.rs.core.Response;
 import javax.ws.rs.core.UriInfo;
 import javax.ws.rs.core.Response.ResponseBuilder;
+
 import javax.xml.namespace.QName;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
+
 import org.apache.hadoop.hbase.HColumnDescriptor;
 import org.apache.hadoop.hbase.HTableDescriptor;
 import org.apache.hadoop.hbase.TableExistsException;
@@ -46,8 +48,7 @@ import org.apache.hadoop.hbase.TableNotFoundException;
 import org.apache.hadoop.hbase.client.HBaseAdmin;
 import org.apache.hadoop.hbase.client.HTable;
 import org.apache.hadoop.hbase.client.HTablePool;
-import org.apache.hadoop.hbase.io.ImmutableBytesWritable;
-import org.apache.hadoop.hbase.stargate.auth.User;
+import org.apache.hadoop.hbase.stargate.User;
 import org.apache.hadoop.hbase.stargate.model.ColumnSchemaModel;
 import org.apache.hadoop.hbase.stargate.model.TableSchemaModel;
 import org.apache.hadoop.hbase.util.Bytes;
@@ -89,31 +90,17 @@ public class SchemaResource implements Constants {
 
   @GET
   @Produces({MIMETYPE_TEXT, MIMETYPE_XML, MIMETYPE_JSON, MIMETYPE_PROTOBUF})
-  public Response get(@Context UriInfo uriInfo) {
+  public Response get(final @Context UriInfo uriInfo) throws IOException {
     if (LOG.isDebugEnabled()) {
       LOG.debug("GET " + uriInfo.getAbsolutePath());
     }
+    if (!servlet.userRequestLimit(user, 1)) {
+      return Response.status(509).build();
+    }
     servlet.getMetrics().incrementRequests(1);
     try {
-      HTableDescriptor htd = getTableSchema();
-      TableSchemaModel model = new TableSchemaModel();
-      model.setName(tableName);
-      for (Map.Entry<ImmutableBytesWritable, ImmutableBytesWritable> e:
-          htd.getValues().entrySet()) {
-        model.addAttribute(Bytes.toString(e.getKey().get()), 
-            Bytes.toString(e.getValue().get()));
-      }
-      for (HColumnDescriptor hcd: htd.getFamilies()) {
-        ColumnSchemaModel columnModel = new ColumnSchemaModel();
-        columnModel.setName(hcd.getNameAsString());
-        for (Map.Entry<ImmutableBytesWritable, ImmutableBytesWritable> e:
-          hcd.getValues().entrySet()) {
-        columnModel.addAttribute(Bytes.toString(e.getKey().get()), 
-          Bytes.toString(e.getValue().get()));
-      }
-        model.addColumnFamily(columnModel);
-      }
-      ResponseBuilder response = Response.ok(model);
+      ResponseBuilder response =
+        Response.ok(new TableSchemaModel(getTableSchema()));
       response.cacheControl(cacheControl);
       return response.build();
     } catch (TableNotFoundException e) {
@@ -124,8 +111,9 @@ public class SchemaResource implements Constants {
     }
   }
 
-  private Response replace(byte[] tableName, TableSchemaModel model,
-      UriInfo uriInfo, HBaseAdmin admin) {
+  private Response replace(final byte[] tableName, 
+      final TableSchemaModel model, final UriInfo uriInfo,
+      final HBaseAdmin admin) {
     try {
       HTableDescriptor htd = new HTableDescriptor(tableName);
       for (Map.Entry<QName,Object> e: model.getAny().entrySet()) {
@@ -155,8 +143,8 @@ public class SchemaResource implements Constants {
     }      
   } 
 
-  private Response update(byte[] tableName, TableSchemaModel model,
-      UriInfo uriInfo, HBaseAdmin admin) {
+  private Response update(final byte[] tableName,final TableSchemaModel model,
+      final UriInfo uriInfo, final HBaseAdmin admin) {
     try {
       HTableDescriptor htd = admin.getTableDescriptor(tableName);
       admin.disableTable(tableName);
@@ -185,8 +173,8 @@ public class SchemaResource implements Constants {
     }
   }
 
-  private Response update(TableSchemaModel model, boolean replace,
-      UriInfo uriInfo) {
+  private Response update(final TableSchemaModel model, final boolean replace,
+      final UriInfo uriInfo) {
     try {
       servlet.invalidateMaxAge(tableName);
       byte[] tableName = Bytes.toBytes(actualTableName);
@@ -204,45 +192,53 @@ public class SchemaResource implements Constants {
 
   @PUT
   @Consumes({MIMETYPE_XML, MIMETYPE_JSON, MIMETYPE_PROTOBUF})
-  public Response put(TableSchemaModel model, @Context UriInfo uriInfo) {
+  public Response put(final TableSchemaModel model, 
+      final @Context UriInfo uriInfo) throws IOException {
     if (LOG.isDebugEnabled()) {
       LOG.debug("PUT " + uriInfo.getAbsolutePath());
     }
-    servlet.getMetrics().incrementRequests(1);
-    // use the name given in the path, but warn if the name on the path and
-    // the name in the schema are different
-    if (model.getName() != tableName) {
-      LOG.warn("table name mismatch: path='" + tableName + "', schema='" +
-        model.getName() + "'");
+    if (!servlet.userRequestLimit(user, 1)) {
+      return Response.status(509).build();
     }
+    servlet.getMetrics().incrementRequests(1);
     return update(model, true, uriInfo);
   }
 
   @POST
   @Consumes({MIMETYPE_XML, MIMETYPE_JSON, MIMETYPE_PROTOBUF})
-  public Response post(TableSchemaModel model, @Context UriInfo uriInfo) {
+  public Response post(final TableSchemaModel model, 
+      final @Context UriInfo uriInfo) throws IOException {
     if (LOG.isDebugEnabled()) {
       LOG.debug("PUT " + uriInfo.getAbsolutePath());
     }
-    servlet.getMetrics().incrementRequests(1);
-    // use the name given in the path, but warn if the name on the path and
-    // the name in the schema are different
-    if (model.getName() != tableName) {
-      LOG.warn("table name mismatch: path='" + tableName + "', schema='" +
-        model.getName() + "'");
+    if (!servlet.userRequestLimit(user, 1)) {
+      return Response.status(509).build();
     }
+    servlet.getMetrics().incrementRequests(1);
     return update(model, false, uriInfo);
   }
 
   @DELETE
-  public Response delete(@Context UriInfo uriInfo) {     
+  public Response delete(final @Context UriInfo uriInfo) throws IOException {
     if (LOG.isDebugEnabled()) {
       LOG.debug("DELETE " + uriInfo.getAbsolutePath());
+    }
+    if (!servlet.userRequestLimit(user, 1)) {
+      return Response.status(509).build();
     }
     servlet.getMetrics().incrementRequests(1);
     try {
       HBaseAdmin admin = new HBaseAdmin(servlet.getConfiguration());
-      admin.disableTable(actualTableName);
+      boolean success = false;
+      for (int i = 0; i < 10; i++) try {
+        admin.disableTable(actualTableName);
+        success = true;
+        break;
+      } catch (IOException e) {
+      }
+      if (!success) {
+        throw new IOException("could not disable table");
+      }
       admin.deleteTable(actualTableName);
       return Response.ok().build();
     } catch (TableNotFoundException e) {
@@ -252,4 +248,5 @@ public class SchemaResource implements Constants {
             Response.Status.SERVICE_UNAVAILABLE);
     }
   }
+
 }
